Updates to DefaultGazeboComponent and docs
The DefaultGazeboComponent now does gazebo interaction in one thread,
and Orocos interaction in the other, so that control computation is done
in the component's updateHook().

Updated docs to make note of the limitations of putting Ops script
inline in XML. Specifically, only `/* ... */` comments can be used
because Gazebo removes all line breaks from parsed XML! Single-line
comments starting with `//` or `#` will comment out the rest of the
script in the `<opsScript>` block, and it will not run.